+++ upgrade/install-2.3.6.php
@@ -1,0 +1,18 @@
+<?php
+/**
+ * Payline module for PrestaShop
+ *
+ * @author    Monext <support@payline.com>
+ * @copyright Monext - http://www.payline.com
+ */
+
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
+function upgrade_module_2_3_6($module)
+{
+    $module->installTab();
+
+    return true;
+}